Attempted to implement PersonSearch and TitleSearch. Current issues:
Pagination incorrect for SearchTitle in TitleController. - Currently no way of counting total entities, needed for pagination. Hardcoding entity count works.

string id in SearchTitle in TitleController should be changed. Currently its called "id" and is not taken from Query as it should be.

SearchPerson in PersonController not working, able to perform correct pagination.
Currently the number of entities here is counted incorrectly, probs from Persons.
The url is looking correct.

added cases for SearchPerson and SearchTitle in SpawnDTO(with pagination).